requirements.txt
boto3==1.34.131
Pillow==11.3.0
